requirements.txt
Flask==3.1.0
pandas==2.2.3
geopandas==1.0.1
requests==2.32.3
requests-cache==1.2.1
folium==0.19.5
plotly==6.0.1
tqdm==4.67.1
shapely==2.0.7
fiona==1.10.1
pyproj==3.7.1
openpyxl==3.1.5
click==8.1.8
gunicorn==23.0.0
python-dotenv==1.1.0
